Simplified and clarified `daysOutOfDate` calculation logic in `Stock.kt`. #ai-message

+++ src/main/java/com/gildedrose/updating/Stock.kt
@@ -22,17 +22,13 @@ class Stock(
     context(TXContext)
     fun loadAndUpdateStockList(now: Instant): Result4k<StockList, StockListLoadingError> {
         val loadedStockList = items.load().onFailure { return it }
-        val daysOutOfDate = loadedStockList.lastModified.daysTo(now, zoneId)
-        if (daysOutOfDate <= 0L) return Success(loadedStockList)
+        val daysOutOfDate = loadedStockList.lastModified.daysTo(now, zoneId).toInt()
+        if (daysOutOfDate <= 0) return Success(loadedStockList)
 
         val updatedItems = loadedStockList.items.map {
-            it.itemUpdate(
-                daysOutOfDate.toInt(),
-                LocalDate.ofInstant(now, zoneId)
-            )
+            it.itemUpdate(daysOutOfDate, LocalDate.ofInstant(now, zoneId))
         }
-        return items.save(loadedStockList.copy(lastModified = now, items = updatedItems
-            ))
+        return items.save(loadedStockList.copy(lastModified = now, items = updatedItems))
     }
 }
